refactor(scripts): Use TypeScript
