+++ package.json
@@ -46,7 +46,7 @@
     "lint": "eslint test/ src/ bin/ lib-src/ scripts/ --cache --ext=js,mjs,mts",
     "lint:fix": "eslint test/ src/ bin/ lib-src/ scripts/ --cache --ext=js,mjs,mts --fix",
     "build": "run-s build:*",
-    "build:regex_data": "node scripts/gen_regex_sets.js",
+    "build:regex_data": "node --enable-source-maps --disable-warning=ExperimentalWarning --experimental-strip-types scripts/gen_regex_sets.mts",
     "build:dts": "tsc -b .",
     "build:engine": "NODE_OPTIONS='--enable-source-maps --disable-warning=ExperimentalWarning --experimental-strip-types' rollup -c ./scripts/rollup.config.mts",
     "build-watch:engine": "npm run build:engine -- --watch",
@@ -57,7 +57,7 @@
     "test:supplemental": "node --enable-source-maps --disable-warning=ExperimentalWarning --experimental-strip-types test/supplemental.mts",
     "test:json": "node --enable-source-maps --disable-warning=ExperimentalWarning --experimental-strip-types test/json/json.mts",
     "coverage": "c8 --reporter=lcov npm run test",
-    "prepublishOnly": "node scripts/tag_version_with_git_hash.js",
+    "prepublishOnly": "node --enable-source-maps --disable-warning=ExperimentalWarning --experimental-strip-types scripts/tag_version_with_git_hash.mts",
     "postpublish": "git reset --hard HEAD"
   },
   "bin": {
